backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.12.4
pydantic-settings==2.1.0
python-multipart==0.0.6

# Google Cloud & AI
google-genai>=1.49.0
google-cloud-aiplatform==1.42.1
google-cloud-storage==2.14.0
googlemaps==4.10.0

# Utilities
pytrends==4.9.2
httpx>=0.28.1
aiofiles==23.2.1
python-dotenv==1.0.0
